+++ src/server/api/forums.ts
@@ -203,6 +203,7 @@ export async function createComment(r) {
     return error(500000, "Internal server error", {});
   }
 }
+
 export const createFile = async (r) => {
   try {
     const file: any = await Files.findOne({
